fix bootstrap to ignore new targets when not present

+++ ansible/bootstrap.sh
@@ -9,7 +9,7 @@ chmod 744 createdfcvolumes.sh
 echo Create FS dirs
 for disk in "$@"; do
     parallel-ssh -h inventory/targets.txt -P 'sudo mkdir -p /dfc/'$disk
-    parallel-ssh -h inventory/new_targets.txt -P 'sudo mkdir -p /dfc/'$disk
+    if [[ -s inventory/new_targets.txt ]]; then parallel-ssh -h inventory/new_targets.txt -P 'sudo mkdir -p /dfc/'$disk; fi
 done
 
 echo Wait 30 sec before creating FS
@@ -19,6 +19,6 @@ echo Create FS
 for disk in "$@"; do
     echo Create XFS on $disk
     parallel-ssh -h inventory/targets.txt -P 'sudo mkfs -t xfs /dev/'$disk
-    parallel-ssh -h inventory/new_targets.txt -P 'sudo mkfs -t xfs /dev/'$disk
+    if [[ -s inventory/new_targets.txt ]]; then parallel-ssh -h inventory/new_targets.txt -P 'sudo mkfs -t xfs /dev/'$disk; fi
 done
